Split project into api and interactive
