+++ internal/products/product_service.go
