+++ internal/products/product_service_test.go
